requirements.txt
# AI & Machine Learning
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Database
mysql-connector-python>=8.0.33

# Web Framework
flask>=2.3.0

# Visualization
matplotlib>=3.7.0

# Optional (jika diperlukan)
# seaborn>=0.12.0
transformers>=4.30.0
sentencepiece
